Bump semver to 0.2, remove version_matches()
semver 0.2 was the first to include a fix to how `VersionReq`
comparisons work with pre-release identifiers. The fix results in, eg,

    version_matches(">= 1.4.0")

not matching for a compiler version `1.8.0-nightly`. For more details,
see https://github.com/steveklabnik/semver/pull/54.

This makes `version_matches()` far less useful, and so it is removed.
It's still possible to get the original intent by using the `Ord` impl
on `semver::Version`. To make this more convenient, `semver::Version` is
exported from this crate.

This is a breaking change.