Improve README example
The README instructs the user to add `rust_c` to the dependencies section, while the following example requires it to be in the build-dependencies section. Since it seems build scripts are the primary use-case for this script, we can reduce friction with this change (it caused me about 5 minutes worth of research).